[CS] move add x to group pages
