+++ app/controllers/actions/ClientAction.scala
@@ -26,7 +26,7 @@ import services.ClientService
 import uk.gov.hmrc.agentmtdidentifiers.model.Arn
 import uk.gov.hmrc.auth.core._
 import uk.gov.hmrc.http.HeaderCarrier
-import views.html.groups.add_groups_to_client.client_not_found
+import views.html.group_member_details.add_groups_to_client.client_not_found
 
 import javax.inject.{Inject, Singleton}
 import scala.concurrent.{ExecutionContext, Future}
